rename rules and separate
