+++ crates/ruff_linter/src/rules/pyupgrade/rules/pep695/use_pep695_generic_function.rs
@@ -1,7 +1,7 @@
 use ruff_diagnostics::{Applicability, Diagnostic, Edit, Fix, FixAvailability, Violation};
 use ruff_macros::{derive_message_formats, ViolationMetadata};
-use ruff_python_ast::{visitor::Visitor, Expr, ExprSubscript};
-use ruff_python_ast::{StmtClassDef, StmtFunctionDef};
+use ruff_python_ast::visitor::Visitor;
+use ruff_python_ast::StmtFunctionDef;
 use ruff_text_size::{Ranged, TextRange};
 
 use crate::checkers::ast::Checker;
@@ -11,13 +11,12 @@ use super::{check_type_vars, in_nested_context, DisplayTypeVars, TypeVarReferenc
 
 /// ## What it does
 ///
-/// Checks for use of standalone type variables and parameter specifications in generic functions
-/// and classes.
+/// Checks for use of standalone type variables and parameter specifications in generic functions.
 ///
 /// ## Why is this bad?
 ///
 /// Special type parameter syntax was introduced in Python 3.12 by [PEP 695] for defining generic
-/// functions and classes. This syntax is easier to read and provides cleaner support for generics.
+/// functions. This syntax is easier to read and provides cleaner support for generics.
 ///
 /// ## Known problems
 ///
@@ -25,10 +24,9 @@ use super::{check_type_vars, in_nested_context, DisplayTypeVars, TypeVarReferenc
 /// `contravariant` keywords used by `TypeVar` variables. As such, rewriting a `TypeVar` variable to
 /// an inline type parameter may change its variance.
 ///
-/// The rule currently skips generic classes with multiple base classes, and skips generic methods
-/// in generic or non-generic classes. It also skips generic functions and classes nested inside of
-/// other functions or classes. Finally, this rule skips type parameters with the `default` argument
-/// introduced in [PEP 696] and implemented in Python 3.13.
+/// The rule currently skips generic functions nested inside of other functions or classes and those
+/// with type parameters containing the `default` argument introduced in [PEP 696] and implemented
+/// in Python 3.13.
 ///
 /// ## Example
 ///
@@ -38,14 +36,14 @@ use super::{check_type_vars, in_nested_context, DisplayTypeVars, TypeVarReferenc
 /// T = TypeVar("T")
 ///
 ///
-/// class GenericClass(Generic[T]):
+/// def generic_function(t: T):
 ///     var: T
 /// ```
 ///
 /// Use instead:
 ///
 /// ```python
-/// class GenericClass[T]:
+/// def generic_function[T](t: T):
 ///     var: T
 /// ```
 ///
@@ -59,31 +57,17 @@ use super::{check_type_vars, in_nested_context, DisplayTypeVars, TypeVarReferenc
 /// [PEP 695]: https://peps.python.org/pep-0695/
 /// [PEP 696]: https://peps.python.org/pep-0696/
 #[derive(ViolationMetadata)]
-pub(crate) struct NonPEP695TypeParameter {
+pub(crate) struct NonPEP695GenericFunction {
     name: String,
-    generic_kind: GenericKind,
 }
 
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-enum GenericKind {
-    GenericClass,
-    GenericFunction,
-}
-
-impl Violation for NonPEP695TypeParameter {
+impl Violation for NonPEP695GenericFunction {
     const FIX_AVAILABILITY: FixAvailability = FixAvailability::Always;
 
     #[derive_message_formats]
     fn message(&self) -> String {
-        let NonPEP695TypeParameter { name, generic_kind } = self;
-        match generic_kind {
-            GenericKind::GenericClass => {
-                format!("Generic class `{name}` uses `Generic` subclass instead of type parameters")
-            }
-            GenericKind::GenericFunction => {
-                format!("Generic function `{name}` should use type parameters")
-            }
-        }
+        let NonPEP695GenericFunction { name } = self;
+        format!("Generic function `{name}` should use type parameters")
     }
 
     fn fix_title(&self) -> Option<String> {
@@ -91,86 +75,7 @@ impl Violation for NonPEP695TypeParameter {
     }
 }
 
-/// UP046
-pub(crate) fn non_pep695_generic_class(checker: &mut Checker, class_def: &StmtClassDef) {
-    // PEP-695 syntax is only available on Python 3.12+
-    if checker.settings.target_version < PythonVersion::Py312 {
-        return;
-    }
-
-    // don't try to handle generic classes inside other functions or classes
-    if in_nested_context(checker) {
-        return;
-    }
-
-    let StmtClassDef {
-        name,
-        type_params,
-        arguments,
-        ..
-    } = class_def;
-
-    // it's a runtime error to mix type_params and Generic, so bail out early if we see existing
-    // type_params
-    if type_params.is_some() {
-        return;
-    }
-
-    let Some(arguments) = arguments.as_ref() else {
-        return;
-    };
-
-    // TODO(brent) only accept a single, Generic argument for now. I think it should be fine to have
-    // other arguments, but this simplifies the fix just to delete the argument list for now
-    let [Expr::Subscript(ExprSubscript {
-        value,
-        slice,
-        range,
-        ..
-    })] = arguments.args.as_ref()
-    else {
-        return;
-    };
-
-    if !checker.semantic().match_typing_expr(value, "Generic") {
-        return;
-    }
-
-    let vars = {
-        let mut visitor = TypeVarReferenceVisitor {
-            vars: vec![],
-            semantic: checker.semantic(),
-        };
-        visitor.visit_expr(slice);
-        visitor.vars
-    };
-
-    let Some(type_vars) = check_type_vars(vars) else {
-        return;
-    };
-
-    // build the fix as a String to avoid removing comments from the entire function body
-    let type_params = DisplayTypeVars {
-        type_vars: &type_vars,
-        source: checker.source(),
-    };
-
-    checker.diagnostics.push(
-        Diagnostic::new(
-            NonPEP695TypeParameter {
-                name: name.to_string(),
-                generic_kind: GenericKind::GenericClass,
-            },
-            *range,
-        )
-        .with_fix(Fix::applicable_edit(
-            Edit::replacement(type_params.to_string(), name.end(), arguments.end()),
-            Applicability::Safe,
-        )),
-    );
-}
-
-/// UP046
+/// UP047
 pub(crate) fn non_pep695_generic_function(checker: &mut Checker, function_def: &StmtFunctionDef) {
     // PEP-695 syntax is only available on Python 3.12+
     if checker.settings.target_version < PythonVersion::Py312 {
@@ -239,9 +144,8 @@ pub(crate) fn non_pep695_generic_function(checker: &mut Checker, function_def: &
 
     checker.diagnostics.push(
         Diagnostic::new(
-            NonPEP695TypeParameter {
+            NonPEP695GenericFunction {
                 name: name.to_string(),
-                generic_kind: GenericKind::GenericFunction,
             },
             TextRange::new(name.start(), parameters.end()),
         )
